mouse location for pathfinding complete
I have follow a tutorial that uses the pathfinding module to make my pathing system. The way to track mouse position is complete with its separate class that I plan to incoporate into the unit class later on. I made some basic sprites for the background and the mouse grid location.