structure class complete
I made the pathfinder class a child class of the object class (originally the unit class) so it can inherit the attributes from the object class. I renamed the unit class to object class since I decided to make the actual class a child of the pathfinder class so units and structures are on the same level. I need made the structure class and unit class and have them load different sprites depending on the class type. I also made a sprite for where the destination of a path is.