collision system working
able to detect when worker class is colliding with resource class. This will be applied to other units in future. The issue was that updating self.rect.centre for the class didn't work because it is the position of the character (object of Object class) that was changing and colliding with the resource. Have 'hitboxes' showing to test this and should be useful for further testing. The normal collision response will like be pushing against the object in with an equal opposite vector to stop it from moving.