sprite for resource + worker add. + system change to spriting
I added the sprite for when the worker has a resource. I also made a big change and was able to change the sprites of the objects by defining the .image within the class initialisation not in a list in the character class. This will also allow me to use the active structure standin as it can be more easily accessed.