y-sort camera attempt
attempted to add a y-sort camera. This makes pygames draw sprites on different layers according to their y position. This currently is done through a custom draw method. By removing the draw methods in the update functions of all other classes, the game loads properly but the layering still doesn't work. There doesn't seem to be any negative effect on the system by this change however so it will stay unless it becomes an issue later.