Y-sort camera and cameraGroup issue fixed
pygame sprite groups can be appended to eachother like python arrays and thus adding all of the sprites to a single cameragroup for rendering allows for the new update to occur. Y-sort camera has also been fixed as before the logic made each sprite sort itself not globally. By putting all the characters in a list this can work and add the fake 3D effect that it brings.