box camera system and keyboard control implemented
have implemented a dead zone camera which moves when the character hits the border of the box to make sure they don't go off screen. This can be implemented as an alternative camera control. I also implement keyboard control to this box which allows for the desired movement with arrow keys. Only thing left is adding keybinds for character centering and the mouse controlled camera system.