constant offset fix
I have fixed the system with a fixed offset which works when zoom_scale is 1 but as soon as the zoom_scale isn't one the offset is noticable. The drawing system works fine with the fixed offset however. A better will need to use a dynamic offset that also counter acts the scaling of the camera when making a path.