collision system problems
when the units collide they lose pathing which is good for some cases like structures so this will need a new class. when worker interacts with resource the mining does not take priority so need to switch the processes around in update or use collisionin the resource collection systems. The units should not be able to push stationary units