selection system implemented
you can now draw a rect using left click mouse and select units that you want to control and only if they are selected will commands go through. Did this but implementing collosion with the character.rects of the objects and adding an attribute to Object 'selected' which gets reset when they are no selected the next grouping. Only need to implement the green border for selected units or only selected units show their hitbox?